chore(cardinal): clean up message stuff  (#694)
