+++ cardinal/testutils/testutils.go
@@ -3,13 +3,17 @@ package testutils
 import (
 	"crypto/ecdsa"
 	"fmt"
+	"reflect"
 	"testing"
 	"time"
 
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/rotisserie/eris"
 
 	"pkg.world.dev/world-engine/cardinal"
+	"pkg.world.dev/world-engine/cardinal/message"
 	"pkg.world.dev/world-engine/cardinal/types"
+	"pkg.world.dev/world-engine/cardinal/types/engine"
 	"pkg.world.dev/world-engine/sign"
 )
 
@@ -84,4 +88,19 @@ func AddTransactionToWorldByAnyTransaction(
 	}
 
 	_, _ = world.AddTransaction(txID, value, tx)
+}
+
+func GetMessage[In any, Out any](wCtx engine.Context) (*message.MessageType[In, Out], error) {
+	var msg message.MessageType[In, Out]
+	msgType := reflect.TypeOf(msg)
+	tempRes, ok := wCtx.GetMessageByType(msgType)
+	if !ok {
+		return nil, eris.Errorf("Could not find %q, Message may not be registered.", msg.Name())
+	}
+	var _ types.Message = &msg
+	res, ok := tempRes.(*message.MessageType[In, Out])
+	if !ok {
+		return &msg, eris.New("wrong type")
+	}
+	return res, nil
 }
